requirements.txt
pandas>=2.0.0
spotipy>=2.23.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0